Major performance upgrades. Started using OpenMP.

+++ src/airfoil.hpp
@@ -5,7 +5,7 @@
 
 class Airfoil {
 public:
-    static std::optional<Airfoil> points_into_panels(const std::vector<Point>& points, const int nPanels);
+    static std::optional<Airfoil> points_into_panels(const std::vector<Point>& points, const size_t nPanels);
     const std::vector<Panel>& panels() const;
     std::vector<Panel>& panels();
 private:
